requirements.txt
# Azure AI Agent - Microsoft Agent Framework with MCP

agent-framework>=1.0.0b260130
python-dotenv>=1.0.0
azure-identity>=1.19.0
